# bookNetwork/src/main/java/com/ajay/bookNetwork/feedback/FeedbackResponse.java
package com.ajay.bookNetwork.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;


}